feat(users): create delete user fn, update prisma instance.
Added the function to delete a user, we do not erase it from the DB
itself, we are doing a "Soft delete" updating the field status as
deleted.

We changed the import of our PrismaClient, now we are using one global
instance, the object is cached the first time the module is imported
subsequent requests return the cached object rather than creating a new
PrismaClient.